The Project itself in all glory
The last version of the code as of now. Agent info scrapping is commented out because the client didn't need it, but the code is there. Additional failsafe checks for a random bug that happened 20 times in 16000 iterations for no reason. Made so that even if the database has duplicates of a person they get removed from the output spreadshit.